@@ -4,13 +4,13 @@
   "license": "MIT",
   "scripts": {
     "start": "cd gui && react-static start",
-    "build": "cd gui && react-static build",
+    "build": "JOBS=2 && cd gui && react-static build",
     "serve": "cd gui && serve dist -p 3000",
     "storybook": "start-storybook -p 9001 -c .storybook",
     "lint": "standard --fix --verbose | snazzy",
     "lint-ci": "standard --verbose",
-    "test": "NODE_ENV=test jest",
-    "test-ci": "NODE_ENV=test yarn jest gui/ --coverage --reporters jest-silent-reporter",
+    "test": "NODE_ENV=test jest --maxWorkers=2",
+    "test-ci": "NODE_ENV=test yarn jest --maxWorkers=2 gui/ --coverage --reporters jest-silent-reporter",
     "test-e2e": "NODE_ENV=test yarn jest e2e/ -i",
     "test-sol": "./internal/ci/truffle_test",
     "watch": "NODE_ENV=test jest --watchAll --notify",
@@ -23,16 +23,24 @@
     "solidity-cborutils": "^1.0.1"
   },
   "devDependencies": {
+    "@babel/core": "^7.2.2",
+    "@babel/plugin-proposal-class-properties": "^7.2.3",
+    "@babel/plugin-proposal-export-namespace-from": "^7.2.0",
+    "@babel/plugin-proposal-throw-expressions": "^7.2.0",
+    "@babel/plugin-transform-member-expression-literals": "^7.0.0",
+    "@babel/plugin-transform-property-literals": "^7.0.0",
+    "@babel/polyfill": "^7.2.5",
+    "@babel/preset-env": "^7.2.3",
+    "@babel/preset-react": "^7.0.0",
+    "@babel/preset-stage-2": "^7.0.0",
+    "@babel/register": "^7.0.0",
     "@material-ui/core": "^3.8.1",
     "@material-ui/icons": "^3.0.1",
     "@storybook/react": "^4.0.9",
     "axios": "^0.18.0",
+    "babel-core": "^7.0.0-bridge.0",
     "babel-eslint": "^10.0.1",
-    "babel-plugin-transform-es3-member-expression-literals": "^6.22.0",
-    "babel-plugin-transform-es3-property-literals": "^6.22.0",
-    "babel-preset-es2015": "^6.24.1",
-    "babel-preset-stage-2": "^6.24.1",
-    "babel-register": "^6.26.0",
+    "babel-jest": "^23.4.2",
     "bignumber.js": "^8.0.1",
     "cbor": "^4.1.1",
     "change-case": "^3.0.2",
@@ -68,15 +76,17 @@
     "prop-types": "^15.6.2",
     "puppeteer": "^1.10.0",
     "query-string": "^6.1.0",
-    "react": "^16.6.3",
+    "react": "^16.7.0",
     "react-copy-to-clipboard": "^5.0.1",
-    "react-dom": "^16.6.3",
-    "react-hot-loader": "^4.3.4",
+    "react-dom": "^16.7.0",
+    "react-hot-loader": "^4.6.3",
     "react-jss": "^8.6.1",
     "react-redux": "^5.0.7",
-    "react-resize-detector": "^3.2.0",
-    "react-router": "^4.3.1",
-    "react-static": "^5.9.12",
+    "react-resize-detector": "^3.4.0",
+    "react-router-dom": "^4.3.1",
+    "react-static": "^6.3.3",
+    "react-static-plugin-jss": "^6.3.0",
+    "react-static-plugin-react-router": "^6.3.1",
     "react-time-ago": "^3.0.3",
     "react-universal-component": "^3.0.0",
     "redux": "^4.0.0",
@@ -106,6 +116,9 @@
     "/solidity",
     "/zos.json"
   ],
+  "resolutions": {
+    "babel-core": "7.0.0-bridge.0"
+  },
   "standard": {
     "parser": "babel-eslint",
     "globals": [
