@@ -1,21 +1,21 @@
-import React, { PureComponent } from 'react'
-import CssBaseline from '@material-ui/core/CssBaseline'
-import { bindActionCreators } from 'redux'
-import { connect } from 'react-redux'
-import { Router, Route, Switch } from 'react-static'
-import { Redirect } from 'react-router'
-import Routes from 'react-static-routes'
-import { hot } from 'react-hot-loader'
+import React from 'react'
+import { Root, Routes } from 'react-static'
 import universal from 'react-universal-component'
-import Loading from 'components/Loading'
-import Private from './Private'
+import { Route, Switch, Redirect } from 'react-router-dom'
+import CssBaseline from '@material-ui/core/CssBaseline'
 import PrivateRoute from './PrivateRoute'
+import Private from './Private'
+import Loading from 'components/Loading'
+import { connect } from 'react-redux'
+import { bindActionCreators } from 'redux'
 
+// Asynchronously load routes that are chunked via code-splitting
+// 'import' as a function must take a string. It can't take a variable.
 const uniOpts = { loading: Loading }
 const SignIn = universal(import('./containers/SignIn'), uniOpts)
 const SignOut = universal(import('./containers/SignOut'), uniOpts)
 
-class Layout extends PureComponent {
+class Layout extends React.Component {
   // Remove the server-side injected CSS.
   componentDidMount () {
     const jssStyles = document.getElementById('jss-server-side')
@@ -28,19 +28,17 @@ class Layout extends PureComponent {
     const { redirectTo } = this.props
 
     return (
-      <React.Fragment>
+      <Root>
         <CssBaseline />
 
-        <Router>
-          <Switch>
-            <Route exact path='/signin' component={SignIn} />
-            <PrivateRoute exact path='/signout' component={SignOut} />
-            {redirectTo && <Redirect to={redirectTo} />}
-            <Route component={Private} />
-            <Routes />
-          </Switch>
-        </Router>
-      </React.Fragment>
+        <Switch>
+          <Route exact path='/signin' component={SignIn} />
+          <PrivateRoute exact path='/signout' component={SignOut} />
+          {redirectTo && <Redirect to={redirectTo} />}
+          <Route component={Private} />
+          <Route render={() => <Routes />} />
+        </Switch>
+      </Root>
     )
   }
 }
@@ -59,4 +57,4 @@ export const ConnectedLayout = connect(
   mapDispatchToProps
 )(Layout)
 
-export default hot(module)(ConnectedLayout)
+export default ConnectedLayout
