@@ -3,7 +3,7 @@ import React from 'react'
 import { Provider } from 'react-redux'
 import { mount } from 'enzyme'
 import { MemoryRouter } from 'react-router'
-import { Switch, Route } from 'react-static'
+import { Switch, Route } from 'react-router-dom'
 import createStore from 'connectors/redux'
 import New from 'containers/Bridges/New'
 import syncFetch from 'test-helpers/syncFetch'
