@@ -3,7 +3,7 @@ import createStore from 'connectors/redux'
 import syncFetch from 'test-helpers/syncFetch'
 import { mount } from 'enzyme'
 import { Provider } from 'react-redux'
-import { Switch, Route } from 'react-static'
+import { Switch, Route } from 'react-router-dom'
 import { MemoryRouter } from 'react-router'
 import SignIn from 'containers/SignIn'
 import fillIn from 'test-helpers/fillIn'
