@@ -1,6 +1,5 @@
 import React, { PureComponent } from 'react'
 import { Provider } from 'react-redux'
-import { hot } from 'react-hot-loader'
 import createStore from 'connectors/redux'
 import Layout from 'Layout'
 import './index.css'
@@ -23,4 +22,4 @@ class App extends PureComponent {
   }
 }
 
-export default hot(module)(App)
+export default (App)
