@@ -4,7 +4,7 @@ import * as formik from 'formik'
 import { withStyles } from '@material-ui/core/styles'
 import Button from '@material-ui/core/Button'
 import { TextField, Grid } from '@material-ui/core'
-import { Prompt } from 'react-static'
+import { Prompt } from 'react-router-dom'
 
 const styles = theme => ({
   card: {
